% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_functions}
\alias{db_functions}
\alias{set_db}
\alias{get_db}
\title{Database Management Functions}
\usage{
set_db(api, driver, ...)

get_db(api)
}
\arguments{
\item{api}{An object representing the API.}

\item{driver}{The driver for the database (e.g., "local", "mongodb").}

\item{...}{Additional arguments to be passed to the database driver
constructor.}
}
\value{
The API object with the database set.
}
\description{
This set of functions provides functionality for managing the
database associated with the API.
}
\details{
\itemize{

\item \code{set_db}: Creates a new database object based on the
specified driver and sets it as the API database. It
dispatches to the appropriate method for constructing the
database object.

\item \code{get_db}: Retrieves the database associated with the API.

}
}
