% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparsers.R
\name{deparsing_functions}
\alias{deparsing_functions}
\alias{deparse_array}
\alias{deparse_datetime}
\alias{deparse_json}
\title{Deparsing Functions}
\usage{
deparse_array(x)

deparse_datetime(x)

deparse_json(x, ...)
}
\arguments{
\item{x}{The data to be converted.}

\item{...}{Additional arguments to be passed to \code{jsonlite::toJSON}.}
}
\description{
This set of functions provides functionality for converting R data
types back to HTTP input parameter strings.
}
\details{
\itemize{

\item \code{deparse_array}: Converts a vector to an array-like string.

\item \code{deparse_datetime}: Converts a datetime object to an STAC
interval-like string.

\item \code{deparse_json}: Converts an R object to a GeoJSON string.

}
}
\seealso{
\code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}} For more details on what can be passed
to \code{...} parameter.
}
