% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{link_functions}
\alias{link_functions}
\alias{update_link}
\alias{root_url}
\alias{self_url}
\alias{parent_url}
\title{Link helper functions}
\usage{
update_link(doc, rel, href, ...)

root_url(api, req)

self_url(api, req, ...)

parent_url(api, req)
}
\arguments{
\item{doc}{The document to which the link will be added.}

\item{rel}{The relationship of the link (e.g., "self", "child").}

\item{href}{The URL of the linked resource.}

\item{...}{Additional attributes to include in the link.}
}
\value{
The updated document with the new link added.
}
\description{
These functions modify links in an document by adding new links or
updating existing ones.
}
\details{
\itemize{

\item \code{add_link}: This function adds a new link to the document by
appending it to the existing links.

\item \code{update_link}: This function updates an existing link in the
document by replacing it with a new link. The existing link is
identified by its relationship (rel). If multiple links with
the same relationship exist, all of them are replaced.

}
}
