% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openstac-package.R
\docType{package}
\name{openstac-package}
\alias{openstac-package}
\alias{openstac}
\title{openstac: STAC API Server}
\description{
Implements STAC and OGC API Features server. Users can connect to MongoDB database to create robust STAC API servers in production environments. With streamlined workflows for generating STAC documents and deploying datasets, users can easily share their geospatial data with others. openstac design facilitates package extension to accommodate additional components of the STAC specification.
}
\section{Purpose}{


The package provides a comprehensive set of functions for building
Spatio-Temporal Asset Catalog (STAC) API implementations. With
\code{openstac}, users can easily deploy geospatial data into a STAC API
service enhancing interoperability with other geotechnologies.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Open-Earth-Monitor/openstac}
  \item Report bugs at \url{https://github.com/Open-Earth-Monitor/openstac/issues}
}

}
\author{
\strong{Maintainer}: Rolf Simoes \email{rolf.simoes@opengeohub.org} (\href{https://orcid.org/0000-0003-0953-4132}{ORCID})

Authors:
\itemize{
  \item Murat Sahin \email{murat.sahin@opengeohub.org} (\href{https://orcid.org/0000-0003-4143-1467}{ORCID})
}

Other contributors:
\itemize{
  \item OpenGeoHub Foundation [copyright holder]
}

}
\keyword{internal}
