% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_handling}
\alias{api_handling}
\alias{api_landing_page}
\alias{api_conformance}
\alias{api_collections}
\alias{api_collection}
\alias{api_items}
\alias{api_item}
\alias{api_search}
\title{Handle API requests}
\usage{
api_landing_page(api, req, res, ...)

api_conformance(api, req, res, ...)

api_collections(api, req, res, ...)

api_collection(api, req, res, collection_id, ...)

api_items(api, req, res, collection_id, limit, bbox, datetime, page, ...)

api_item(api, req, res, collection_id, item_id, ...)

api_search(
  api,
  req,
  res,
  limit,
  bbox,
  datetime,
  intersects,
  ids,
  collections,
  page,
  ...
)
}
\arguments{
\item{api}{An object representing the API. This object is typically
created using either the \code{create_stac} or \code{create_ogcapi} function.}

\item{req}{The request object from the \code{plumber} package, containing
information about the HTTP request made to the API endpoint.}

\item{res}{The response object from the \code{plumber} package, used to
construct and send the HTTP response back to the client making
the request.}

\item{...}{Additional arguments to be passed to the method-specific
functions.}

\item{collection_id}{The identifier of the collection. This parameter
specifies which collection the request is targeting.}
}
\description{
These functions are responsible for handling requests of the API
endpoint. It interfaces HTTP requests from \code{plumber} and uses the
\code{api}, the \code{req}, and the \code{res} objects to prepare a response to the
request by dispatching to specific API implementations.
}
\details{
\itemize{

\item \code{api_landing_page}: handles the STAC \code{/} endpoint

\item \code{api_conformance}: handles the STAC \verb{/conformance} endpoint

\item \code{api_collections}: handles the STAC \verb{/collections} endpoint

\item \code{api_collection}: handles the STAC
\verb{/collection/\\\{collection_id\\\}} endpoint

\item \code{api_items}: handles the STAC
\verb{/collection/\\\{collection_id\\\}/items} endpoint

\item \code{api_item}: handles the STAC
\verb{/collection/\\\{collection_id\\\}/item/\\\{item_id\\\}} endpoint

\item \code{api_search}: handles the STAC \verb{/search} endpoint

}
}
\references{
For more information about the STAC specification,
see: \url{https://stacspec.org/}
For more information about the OGC API specification,
see: \url{http://www.opengis.net/doc/IS/ogcapi-features-1/1.0}
}
\seealso{
\code{\link{create_stac}}, \code{\link{create_ogcapi}}
Functions for creating STAC and OGC API objects, respectively.
}
