% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{api_helpers}
\alias{api_helpers}
\alias{api_cors_handler}
\alias{api_error_handler}
\alias{api_stop}
\alias{api_stopifnot}
\alias{get_host}
\alias{get_method}
\title{API helper functions}
\usage{
api_cors_handler(req, res, origin = "*", methods = "*")

api_error_handler(req, res, err)

api_stop(status, ...)

api_stopifnot(expr, status, ...)

get_host(req)

get_method(req)
}
\arguments{
\item{req}{The request object from the \code{plumber} package, containing
information about the HTTP request made to the API endpoint.}

\item{res}{The response object from the \code{plumber} package, used to
construct and send the HTTP response back to the client making
the request.}

\item{origin}{The value to set for the 'Access-Control-Allow-Origin'
header in CORS requests. Defaults to '*'.}

\item{methods}{The value to set for the 'Access-Control-Allow-Methods'
header in CORS requests. Defaults to '*'.}

\item{err}{The error object containing information about the
encountered error. If the error is thrown by \code{api_stopifnot} or
\code{api_stop} functions, this object has 'status' and 'message' fields
that are used to produce the HTTP response error.}

\item{status}{The HTTP status code to set for the response. This just
works if the \code{api_error_handler} function is handling errors in
\code{plumber}.}

\item{...}{Additional arguments to be passed to error handling functions.}

\item{expr}{The expression to evaluate. If the expression evaluates
to FALSE, an error will be raised.}
}
\description{
This set of functions provides HTTP CORS (Cross-Origin Resource Sharing)
support and error handling for the API.
}
\details{
\itemize{

\item \code{api_cors_handler}: HTTP CORS support. Typically called from a
\code{plumber} filter to manage CORS.

\item \code{api_error_handler}: Error handling function to be provided to
\code{plumber} router to construct HTTP errors in a standardized way.
Use \code{plumber::pr_set_error()} function to set the error handler
in \code{plumber}.

\item \code{api_stop}: Throws an error and set the HTTP status code and
error message to be returned to the user by the \code{api_error_handler}
function.

\item \code{api_stopifnot}: Throws an error if the provided expression
is evaluated as \code{FALSE}.

\item \code{get_host}: Get the API host address from an \code{req} object.

\item \code{get_method}: Get the HTTP method from an \code{req} object.

}
}
\references{
The code for \code{api_cors_handler} was based on a issue discussion post
on CORS support in plumber at
\verb{https://github.com/rstudio/plumber/issues/66#issuecomment-418660334}
}
\seealso{
\code{\link[plumber:pr_set_error]{plumber::pr_set_error()}}: Function to set error handler in \code{plumber}.
}
