% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openstac-package.R
\docType{package}
\name{openstac-package}
\alias{openstac-package}
\alias{openstac}
\title{openstac: R STAC Server}
\description{
This package implements an STAC API server for R language.
}
\section{Purpose}{


The package provides a comprehensive set of functions for building
Spatio-Temporal Asset Catalog (STAC) API implementations. With
\code{openstac}, users can easily deploy geospatial data into a STAC API
service enhancing interoperability with other geotechnologies.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Open-Earth-Monitor/openstac}
  \item Report bugs at \url{https://github.com/Open-Earth-Monitor/openstac/issues}
}

}
\author{
\strong{Maintainer}: Rolf Simoes \email{rolf.simoes@opengeohub.org} (\href{https://orcid.org/0000-0003-0953-4132}{ORCID})

Authors:
\itemize{
  \item Murat Sahin \email{murat.sahin@opengeohub.org} (\href{https://orcid.org/0000-0003-4143-1467}{ORCID})
}

Other contributors:
\itemize{
  \item OpenGeoHub Foundation [copyright holder]
}

}
\keyword{internal}
